Develop Numba extensions for NumpyArray and ListOffsetArray (#3)
* First check-in to establish branch for Numba development.

* Directory structure for Numba, including a home for cpu-kernels.so.

* Successfully unboxing.

* Boxing also works.

* Added test for boxing and unboxing Numba.

* Implemented and tested lowered NumpyArray.__getitem__.

* Tested NumpyArray.__len__.

* Name tests after the PR they were created for.

* [skip ci] Starting ListOffsetArray for Numba.

* Only run Numba tests if you can install Numba.

* Actually import pytest.

* Guard against missing dummy1.

* [skip ci] setting up dummy1 as a method on NumpyArray

* Successfully calling dummy1 from a lowered function.

* Get members from Content classes.

* Unboxed ListOffsetsArray; now just have to box it.

* Boxing and unboxing ListOffsetArrays.

* After careful study, I'm sure this is correct reference-counting.

* Still failing for Windows/MacOS Python 2.7; maybe it's the Numba version?

* But it has to compile.

* It's not the Numba version; try one one thing before giving up on Python 2.7.

* Drop support for Numba on {MacOS, Windows} Python 2.7.

* Turn ListOffsetArray's offsets and content into properties.

* Implemented lowered ListOffsetArray.__len__.

* ListOffsetArray.__len__ for the pybind objects, too.

* ListOffsetArray.__getitem__ lowered for integer arguments.

* [skip ci] working on ListOffsetArray.__getitem__(slice)

* [skip ci] Much closer.

* Slicing ListOffsetArrays works.

+++ include/awkward/util.h
@@ -17,7 +17,7 @@
 #endif
 
 namespace awkward {
-  typedef int64_t AtType;
+  typedef int32_t AtType;
   typedef int32_t IndexType;
   typedef uint8_t byte;
